Pass `--test=rspec` to `bundle gem`
The generator expects to patch RSpec files, so specifying anything other
than `rspec` as the test framework would break generation. Therefore, it
makes sense to specify it ourselves, and save the user the time of
having to select it.

By doing this, we can also remove the patch to add the gem, since
Bundler will handle that for us.